Add files via upload
I did a Data Analysis of a user top 50 songd from spotify. I imported the dataset and then analyzed the average Energy and Length of the first 10 songs. 
Then, I find the total length of songs grouped by genre from top to bottom. I then find the artist name with the most number of tracks in one genre.
Finally, I printed the data of the tracks created by the artist.